Added solution to lesson-6
lesson-6 is about using modules. In this lesson we created
two files, the first one is for the function to be exported and
the second one is for the one that would call the module. This
is also an asynchronous exercise.

The filter module filters a directory of files that match
an specified extension and returns it to the calling module as
an array.

The calling module then prints out the files one by one.